requirements.txt
langchain_core
langchain-google-genai 
docling
python-dotenv